fix: Check higher spaces if blocked for any reason
This is a bug that did sneak in with the introduction of the
PathBlocked enum. If a PacketNumberSpace is blocked for any
reason (anti-amplification, congestion or pacing), we still need to be
able to send tail-loss probes on higher PacketNumberSpaces. Otherwise
those spaces could end up in a deadlock too.