fix: prevent the WeakConnectionHandle from keeping the entire allocation for a ConnectionInner alive
ConnectionInner is quite large, so before the weak handle would keep a large
allocation alive. Now it just keeps a tiny allocation alive, consisting of
a pointer and 2 refcounts.

The downside of this approach is that it does a tiny allocation in fn weak_handle.
The alternative would be to do the double arc inside Connection, which would
make every connection access a double dereference, but prevent this
allocation.